Music update

+++ include/game/health.h
@@ -15,7 +15,7 @@ struct render_node_t;
 typedef struct health_t
 {
     WINDOW *window;
-    uint8_t health;
+    uint16_t health;
     uint16_t max_health;
     struct render_node_t* health_node;
 } health_t;
